Clean some dependency in coin-framework
