+++ libs/ledger-live-common/src/bridge/jsHelpers.ts
@@ -5,6 +5,15 @@ import { log } from "@ledgerhq/logs";
 import { WrongDeviceForAccount } from "@ledgerhq/errors";
 import Transport from "@ledgerhq/hw-transport";
 import {
+  AccountShapeInfo,
+  AccountUpdater,
+  GetAccountShape,
+  IterateResult,
+  IterateResultBuilder,
+  mergeOps,
+  sameOp,
+} from "@ledgerhq/coin-framework/lib/bridge/jsHelpers";
+import {
   getSeedIdentifierDerivation,
   getDerivationModesForCurrency,
   getDerivationScheme,
@@ -27,126 +36,31 @@ import {
   recalculateAccountBalanceHistories,
   encodeAccountId,
 } from "../account";
-import { FreshAddressIndexInvalid, UnsupportedDerivation } from "../errors";
+import {
+  FreshAddressIndexInvalid,
+  UnsupportedDerivation,
+} from "@ledgerhq/coin-framework/lib/errors";
 import getAddress from "../hw/getAddress";
 import type { Result, GetAddressOptions } from "../hw/getAddress/types";
 import { withDevice } from "../hw/deviceAccess";
-import type { CryptoCurrency } from "@ledgerhq/types-cryptoassets";
 import type {
   Account,
   AccountBridge,
   CurrencyBridge,
-  Operation,
   ProtoNFT,
   ScanAccountEvent,
-  SyncConfig,
 } from "@ledgerhq/types-live";
 
-// Customize the way to iterate on the keychain derivation
-type IterateResult = ({
-  transport,
-  index,
-  derivationsCache,
-  derivationScheme,
-  derivationMode,
-  currency,
-}: {
-  transport: Transport;
-  index: number;
-  derivationsCache: Record<string, unknown>;
-  derivationScheme: string;
-  derivationMode: DerivationMode;
-  currency: CryptoCurrency;
-}) => Promise<Result | null>;
-
-export type IterateResultBuilder = ({
-  result, // derivation on the "root" of the derivation
-  derivationMode, // identify the current derivation scheme
-  derivationScheme,
-}: {
-  result: Result;
-  derivationMode: DerivationMode;
-  derivationScheme: string;
-}) => Promise<IterateResult>;
-
-export type GetAccountShapeArg0 = {
-  currency: CryptoCurrency;
-  address: string;
-  index: number;
-  initialAccount?: Account;
-  derivationPath: string;
-  derivationMode: DerivationMode;
-  transport?: Transport;
-  rest?: any;
+//-- Export from coin-framework
+export {
+  AccountShapeInfo as GetAccountShapeArg0,
+  AccountUpdater,
+  GetAccountShape,
+  IterateResult,
+  IterateResultBuilder,
+  mergeOps,
+  sameOp,
 };
-
-export type GetAccountShape = (
-  arg0: GetAccountShapeArg0,
-  arg1: SyncConfig
-) => Promise<Partial<Account>>;
-type AccountUpdater = (arg0: Account) => Account;
-
-// compare that two dates are roughly the same date in order to update the case it would have drastically changed
-const sameDate = (a, b) => Math.abs(a - b) < 1000 * 60 * 30;
-
-// an operation is relatively immutable, however we saw that sometimes it can temporarily change due to reorg,..
-export const sameOp = (a: Operation, b: Operation): boolean =>
-  a === b ||
-  (a.id === b.id && // hash, accountId, type are in id
-    (a.fee ? a.fee.isEqualTo(b.fee) : a.fee === b.fee) &&
-    (a.value ? a.value.isEqualTo(b.value) : a.value === b.value) &&
-    a.nftOperations?.length === b.nftOperations?.length &&
-    sameDate(a.date, b.date) &&
-    a.blockHeight === b.blockHeight &&
-    isEqual(a.senders, b.senders) &&
-    isEqual(a.recipients, b.recipients));
-// efficiently prepend newFetched operations to existing operations
-export function mergeOps( // existing operations. sorted (newer to older). deduped.
-  existing: Operation[], // new fetched operations. not sorted. not deduped. time is allowed to overlap inside existing.
-  newFetched: Operation[]
-): // return a list of operations, deduped and sorted from newer to older
-Operation[] {
-  // there is new fetched
-  if (newFetched.length === 0) return existing;
-  // efficient lookup map of id.
-  const existingIds = {};
-
-  for (const o of existing) {
-    existingIds[o.id] = o;
-  }
-
-  // only keep the newFetched that are not in existing. this array will be mutated
-  let newOps = newFetched
-    .filter((o) => !existingIds[o.id] || !sameOp(existingIds[o.id], o))
-    .sort((a, b) => b.date.valueOf() - a.date.valueOf());
-
-  // Deduplicate new ops to guarantee operations don't have dups
-  const newOpsIds = {};
-  newOps.forEach((op) => {
-    newOpsIds[op.id] = op;
-  });
-  newOps = Object.values(newOpsIds);
-
-  // return existing when there is no real new operations
-  if (newOps.length === 0) return existing;
-  // edge case, existing can be empty. return the sorted list.
-  if (existing.length === 0) return newOps;
-  // building up merging the ops
-  const all: Operation[] = [];
-
-  for (const o of existing) {
-    // prepend all the new ops that have higher date
-    while (newOps.length > 0 && newOps[0].date >= o.date) {
-      all.push(newOps.shift() as Operation);
-    }
-
-    if (!newOpsIds[o.id]) {
-      all.push(o);
-    }
-  }
-
-  return all;
-}
 
 export const mergeNfts = (
   oldNfts: ProtoNFT[],
