+++ libs/ledger-live-common/src/errors.ts
@@ -1,13 +1,53 @@
 import { createCustomErrorClass } from "@ledgerhq/errors";
+// TODO we need to migrate in all errors that are in @ledgerhq/errors
+// but only make sense to live-common to not pollute ledgerjs
+export const ConnectAppTimeout = createCustomErrorClass("ConnectAppTimeout");
+export const ConnectManagerTimeout = createCustomErrorClass(
+  "ConnectManagerTimeout"
+);
+export const GetAppAndVersionUnsupportedFormat = createCustomErrorClass(
+  "GetAppAndVersionUnsupportedFormat"
+);
+export const AccountNeedResync = createCustomErrorClass("AccountNeedResync");
 
-export * from "@ledgerhq/coin-framework/lib/errors";
-export * from "./families/polkadot/errors";
-export * from "./families/stellar/errors";
-export * from "./families/solana/errors";
-export * from "./families/cardano/errors";
-export * from "./families/near/errors";
+export const LatestFirmwareVersionRequired = createCustomErrorClass(
+  "LatestFirmwareVersionRequired"
+);
 
-// Tron Error
+export const FeeEstimationFailed = createCustomErrorClass(
+  "FeeEstimationFailed"
+);
+export const LowerThanMinimumRelayFee = createCustomErrorClass(
+  "LowerThanMinimumRelayFee"
+);
+export const TransactionRefusedOnDevice = createCustomErrorClass(
+  "TransactionRefusedOnDevice"
+);
+
+export const LanguageInstallRefusedOnDevice = createCustomErrorClass(
+  "LanguageInstallRefusedOnDevice"
+);
+
+export const ImageLoadRefusedOnDevice = createCustomErrorClass(
+  "ImageLoadRefusedOnDevice"
+);
+
+export const ImageDoesNotExistOnDevice = createCustomErrorClass(
+  "ImageDoesNotExistOnDevice"
+);
+
+export const ImageCommitRefusedOnDevice = createCustomErrorClass(
+  "ImageCommitRefusedOnDevice"
+);
+
+export const LanguageInstallTimeout = createCustomErrorClass(
+  "LanguageInstallTimeout"
+);
+
+export const DeviceNotOnboarded = createCustomErrorClass("DeviceNotOnboarded");
+export const InvalidAddressBecauseAlreadyDelegated = createCustomErrorClass(
+  "InvalidAddressBecauseAlreadyDelegated"
+);
 export const TronNoFrozenForBandwidth = createCustomErrorClass(
   "TronNoFrozenForBandwidth"
 );
@@ -41,8 +81,10 @@ export const TronTransactionExpired = createCustomErrorClass(
 export const TronNotEnoughEnergy = createCustomErrorClass(
   "TronNotEnoughEnergy"
 );
-
-// Cosmos Error
+export const AccountAwaitingSendPendingOperations = createCustomErrorClass(
+  "AccountAwaitingSendPendingOperations"
+);
+export const SourceHasMultiSign = createCustomErrorClass("SourceHasMultiSign");
 export const CosmosRedelegationInProgress = createCustomErrorClass(
   "CosmosRedelegationInProgress"
 );
@@ -58,6 +100,10 @@ export const CosmosTooManyValidators = createCustomErrorClass(
 export const NotEnoughDelegationBalance = createCustomErrorClass(
   "NotEnoughDelegationBalance"
 );
+export const RPCHostRequired = createCustomErrorClass("RPCHostRequired");
+export const RPCHostInvalid = createCustomErrorClass("RPCHostInvalid");
+export const RPCUserRequired = createCustomErrorClass("RPCUserRequired");
+export const RPCPassRequired = createCustomErrorClass("RPCPassRequired");
 // Note : info of this code can be found here :
 // https://github.com/cosmos/cosmos-sdk/blob/v0.40.0-rc3/types/errors/errors.go#L16
 export const CosmosBroadcastError = {
@@ -98,8 +144,6 @@ export const CosmosBroadcastError = {
   "35": createCustomErrorClass("CosmosBroadcastLogic"),
   "36": createCustomErrorClass("CosmosBroadcastConflict"),
 };
-
-// SatStack
 export const SatStackVersionTooOld = createCustomErrorClass(
   "SatStackVersionTooOld"
 );
@@ -110,8 +154,77 @@ export const SatStackStillSyncing = createCustomErrorClass(
 export const SatStackDescriptorNotImported = createCustomErrorClass(
   "SatStackDescriptorNotImported"
 );
+export const SwapNoAvailableProviders = createCustomErrorClass(
+  "SwapNoAvailableProviders"
+);
+export const NoSuchAppOnProvider = createCustomErrorClass(
+  "NoSuchAppOnProvider"
+);
+export const SwapExchangeRateAmountTooLow = createCustomErrorClass(
+  "SwapExchangeRateAmountTooLow"
+);
+export const SwapExchangeRateAmountTooHigh = createCustomErrorClass(
+  "SwapExchangeRateAmountTooHigh"
+);
 
-// Algorand
+export const SwapCheckKYCStatusFailed = createCustomErrorClass(
+  "SwapCheckKYCStatusFailed"
+);
+
+export const SwapSubmitKYCFailed = createCustomErrorClass(
+  "SwapSubmitKYCFailed"
+);
+
+export const SwapGenericAPIError = createCustomErrorClass(
+  "SwapGenericAPIError"
+);
+
+export const JSONRPCResponseError = createCustomErrorClass(
+  "JSONRPCResponseError"
+);
+export const JSONDecodeError = createCustomErrorClass("JSONDecodeError");
+export const NoIPHeaderError = createCustomErrorClass("NoIPHeaderError");
+export const CurrencyNotSupportedError = createCustomErrorClass(
+  "CurrencyNotSupportedError"
+);
+export const CurrencyDisabledError = createCustomErrorClass(
+  "CurrencyDisabledError"
+);
+export const CurrencyDisabledAsInputError = createCustomErrorClass(
+  "CurrencyDisabledAsInputError"
+);
+export const CurrencyDisabledAsOutputError = createCustomErrorClass(
+  "CurrencyDisabledAsOutputError"
+);
+export const CurrencyNotSupportedByProviderError = createCustomErrorClass(
+  "CurrencyNotSupportedByProviderError"
+);
+export const TradeMethodNotSupportedError = createCustomErrorClass(
+  "TradeMethodNotSupportedError"
+);
+export const UnexpectedError = createCustomErrorClass("UnexpectedError");
+export const NotImplementedError = createCustomErrorClass(
+  "NotImplementedError"
+);
+export const ValidationError = createCustomErrorClass("ValidationError");
+export const AccessDeniedError = createCustomErrorClass("AccessDeniedError");
+
 export const AlgorandASANotOptInInRecipient = createCustomErrorClass(
   "AlgorandASANotOptInInRecipient"
 );
+export const CompoundLowerAllowanceOfActiveAccountError =
+  createCustomErrorClass("CompoundLowerAllowanceOfActiveAccountError");
+export const OutdatedApp = createCustomErrorClass("OutdatedApp");
+
+export const BluetoothNotSupportedError = createCustomErrorClass(
+  "FwUpdateBluetoothNotSupported"
+);
+
+export const EConnResetError = createCustomErrorClass("EConnReset");
+
+export * from "@ledgerhq/coin-framework/lib/errors";
+export * from "./families/polkadot/errors";
+export * from "./families/stellar/errors";
+export * from "./families/solana/errors";
+export * from "./families/cardano/errors";
+export * from "./families/near/errors";
